+++ ChangeLog.rst
@@ -2,6 +2,7 @@
 ==================
 
  * Added timeout
+ * Added socketio client
 
 0.1.0 - 17 Dec 2011
 ===================